+++ src/schemas/auth_schemas.py
@@ -5,7 +5,6 @@ from pydantic import BaseModel, EmailStr
 
 class Token(BaseModel):
     access_token: str
-    refresh_token: str
     token_type: str
 
 class TokenData(BaseModel):
